crossbuild: Add mips platform + possibility to choose the golang builder version

+++ cmd/crossbuild.go
@@ -25,12 +25,9 @@ import (
 )
 
 var (
-	dockerBuilderImageName    = "prom/golang-builder"
-	dockerMainBuilderImage    = fmt.Sprintf("%s:main", dockerBuilderImageName)
-	dockerARMBuilderImage     = fmt.Sprintf("%s:arm", dockerBuilderImageName)
-	dockerPowerPCBuilderImage = fmt.Sprintf("%s:powerpc", dockerBuilderImageName)
-	//dockerMIPSBuilderImage       = fmt.Sprintf("%s:mips", dockerBuilderImageName)
+	dockerBuilderImageName = "prom/golang-builder"
 
+	defaultGoVersion     = "1.5.3"
 	defaultMainPlatforms = []string{
 		"linux/amd64", "linux/386", "darwin/amd64", "darwin/386", "windows/amd64", "windows/386",
 		"freebsd/amd64", "freebsd/386", "openbsd/amd64", "openbsd/386", "netbsd/amd64", "netbsd/386",
@@ -42,9 +39,10 @@ var (
 	defaultPowerPCPlatforms = []string{
 		"linux/ppc64", "linux/ppc64le",
 	}
-	/*defaultMIPSPlatforms = []string{
-		"linux/mips", "linux/mipsel",
-	}*/
+	// 1.6 "linux/mips64", "linux/mips64le",
+	defaultMIPSPlatforms = []string{
+		"",
+	}
 )
 
 // crossbuildCmd represents the crossbuild command
@@ -65,9 +63,9 @@ var crossbuildCmd = &cobra.Command{
 		if !viper.IsSet("crossbuild.platforms.powerpc") {
 			viper.Set("crossbuild.platforms.powerpc", defaultPowerPCPlatforms)
 		}
-		//if !viper.IsSet("crossbuild.platforms.mips") {
-		//	viper.SetDefault("crossbuild.platforms.mips", defaultMIPSPlatforms)
-		//}
+		if !viper.IsSet("crossbuild.platforms.mips") {
+			viper.Set("crossbuild.platforms.mips", defaultMIPSPlatforms)
+		}
 	},
 }
 
@@ -75,18 +73,21 @@ var crossbuildCmd = &cobra.Command{
 func init() {
 	Promu.AddCommand(crossbuildCmd)
 
+	crossbuildCmd.Flags().String("go", "", "Golang builder version to use")
 	crossbuildCmd.Flags().String("repo-path", "", "Project repository path")
 	crossbuildCmd.Flags().String("main", "", "Main platforms to build")
 	crossbuildCmd.Flags().String("arm", "", "ARM platforms to build")
 	crossbuildCmd.Flags().String("powerpc", "", "PowerPC platforms to build")
-	//crossbuildCmd.Flags().String("mips", "", "MIPS platforms to build")
+	crossbuildCmd.Flags().String("mips", "", "MIPS platforms to build")
 
+	viper.BindPFlag("go", crossbuildCmd.Flags().Lookup("go"))
 	viper.BindPFlag("repository.path", crossbuildCmd.Flags().Lookup("repo-path"))
 	viper.BindPFlag("crossbuild.platforms.main", crossbuildCmd.Flags().Lookup("main"))
 	viper.BindPFlag("crossbuild.platforms.arm", crossbuildCmd.Flags().Lookup("arm"))
 	viper.BindPFlag("crossbuild.platforms.powerpc", crossbuildCmd.Flags().Lookup("powerpc"))
-	//viper.BindPFlag("crossbuild.platforms.mips", crossbuildCmd.Flags().Lookup("mips"))
+	viper.BindPFlag("crossbuild.platforms.mips", crossbuildCmd.Flags().Lookup("mips"))
 
+	viper.SetDefault("go", defaultGoVersion)
 	// Current bug in viper: SeDefault doesn't work with nested key
 	//viper.SetDefault("crossbuild.platforms.main", defaultMainPlatforms)
 	//viper.SetDefault("crossbuild.platforms.arm", defaultARMPlatforms)
@@ -103,34 +104,40 @@ func runCrossbuild() {
 	}
 
 	var (
+		goVersion        = viper.GetString("go")
 		repoPath         = viper.GetString("repository.path")
 		mainPlatforms    = viper.GetStringSlice("crossbuild.platforms.main")
 		ARMPlatforms     = viper.GetStringSlice("crossbuild.platforms.arm")
 		powerPCPlatforms = viper.GetStringSlice("crossbuild.platforms.powerpc")
-		//MIPSPlatforms    = viper.GetStringSlice("crossbuild.platforms.mips")
+		MIPSPlatforms    = viper.GetStringSlice("crossbuild.platforms.mips")
+
+		dockerMainBuilderImage    = fmt.Sprintf("%s:%s-main", dockerBuilderImageName, goVersion)
+		dockerARMBuilderImage     = fmt.Sprintf("%s:%s-arm", dockerBuilderImageName, goVersion)
+		dockerPowerPCBuilderImage = fmt.Sprintf("%s:%s-powerpc", dockerBuilderImageName, goVersion)
+		dockerMIPSBuilderImage    = fmt.Sprintf("%s:%s-mips", dockerBuilderImageName, goVersion)
 	)
 
-	if len(mainPlatforms) > 0 {
+	if mainPlatformsParam := strings.Join(mainPlatforms[:], " "); mainPlatformsParam != "" {
 		fmt.Println("> running the main builder docker image")
 		sh("docker run --rm -t -v $PWD:/app",
-			dockerMainBuilderImage, "-i", repoPath, "-p", q(strings.Join(mainPlatforms[:], " ")))
+			dockerMainBuilderImage, "-i", repoPath, "-p", q(mainPlatformsParam))
 	}
 
-	if len(ARMPlatforms) > 0 {
+	if ARMPlatformsParam := strings.Join(ARMPlatforms[:], " "); ARMPlatformsParam != "" {
 		fmt.Println("> running the ARM builder docker image")
 		sh("docker run --rm -t -v $PWD:/app",
-			dockerARMBuilderImage, "-i", repoPath, "-p", q(strings.Join(ARMPlatforms[:], " ")))
+			dockerARMBuilderImage, "-i", repoPath, "-p", q(ARMPlatformsParam))
 	}
 
-	if len(powerPCPlatforms) > 0 {
+	if powerPCPlatformsParam := strings.Join(powerPCPlatforms[:], " "); powerPCPlatformsParam != "" {
 		fmt.Println("> running the PowerPC builder docker image")
 		sh("docker run --rm -t -v $PWD:/app",
-			dockerPowerPCBuilderImage, "-i", repoPath, "-p", q(strings.Join(powerPCPlatforms[:], " ")))
+			dockerPowerPCBuilderImage, "-i", repoPath, "-p", q(powerPCPlatformsParam))
 	}
 
-	/*if len(MIPSPlatforms) > 0 {
+	if MIPSPlatformsParam := strings.Join(MIPSPlatforms[:], " "); MIPSPlatformsParam != "" {
 		fmt.Println("> running the MIPS builder docker image")
 		sh("docker run --rm -t -v $PWD:/app",
-			dockerMIPSBuilderImage, "-i", repoPath, "-p", q(strings.Join(MIPSPlatforms[:], " ")))
-	}*/
+			dockerMIPSBuilderImage, "-i", repoPath, "-p", q(MIPSPlatformsParam))
+	}
 }
